Added new help string

+++ src/cuda_util/cuda_scheduler.cpp
@@ -13,14 +13,14 @@ namespace fmm {
       , slots_per_cuda_stream(1) // Slots (queue per stream) is currently deactived
       , number_slots(number_cuda_streams_managed * slots_per_cuda_stream) {
         // Determine what the scheduler has to manage
-        const int total_worker_count = hpx::get_os_thread_count();
-        const int worker_id = hpx::get_worker_thread_num();
+        const size_t total_worker_count = hpx::get_os_thread_count();
+        const size_t worker_id = hpx::get_worker_thread_num();
         const size_t streams_per_locality = opts.cuda_streams_per_locality;
         const size_t streams_per_gpu = opts.cuda_streams_per_gpu;
         size_t gpu_count = streams_per_locality / streams_per_gpu;
         if (streams_per_locality % streams_per_gpu != 0)
             gpu_count++;
-        int number_of_streams_managed = streams_per_locality / total_worker_count;
+        size_t number_of_streams_managed = streams_per_locality / total_worker_count;
         const size_t remaining_streams = streams_per_locality % total_worker_count;
         size_t offset = 0;
         if (remaining_streams != 0) {
@@ -28,14 +28,14 @@ namespace fmm {
                 offset = 1;
         }
 
-        const int accumulated_offset = worker_id < remaining_streams ? worker_id : remaining_streams;
-        const int worker_stream_id = worker_id * number_of_streams_managed +
+        const size_t accumulated_offset = worker_id < remaining_streams ? worker_id : remaining_streams;
+        const size_t worker_stream_id = worker_id * number_of_streams_managed +
         accumulated_offset;
-        const int gpu_id = (worker_stream_id) / streams_per_gpu;
+        const size_t gpu_id = (worker_stream_id) / streams_per_gpu;
         std::cout << "Worker " << worker_id << " uses gpu " << gpu_id << std::endl;
         number_of_streams_managed += offset;
         number_cuda_streams_managed = number_of_streams_managed;
-	number_slots = number_cuda_streams_managed * slots_per_cuda_stream; 
+	number_slots = number_cuda_streams_managed * slots_per_cuda_stream;
 
         local_expansions_slots = std::vector<struct_of_array_data<expansion, real, 20, ENTRIES,
             SOA_PADDING, std::vector<real, cuda_pinned_allocator<real>>>>(number_slots);
@@ -74,7 +74,7 @@ namespace fmm {
         //stream_interfaces = std::vector<util::cuda_helper>(number_cuda_streams_managed);
         stream_interfaces.reserve(number_cuda_streams_managed);
         for (kernel_device_enviroment& env : kernel_device_enviroments) {
-            const int worker_gpu_id = (worker_stream_id + local_stream_id) / streams_per_gpu;
+            const size_t worker_gpu_id = (worker_stream_id + local_stream_id) / streams_per_gpu;
             util::cuda_helper::cuda_error(cudaSetDevice(worker_gpu_id));
             stream_interfaces.emplace_back(worker_gpu_id);
